# Conflicts:
#	ui/model/wallet_model.h
#	wallet/client/extensions/offers_board/swap_offers_board.cpp
#	wallet/client/extensions/offers_board/swap_offers_board.h
#	wallet/client/wallet_client.cpp
#	wallet/client/wallet_client.h
#	wallet/core/common.cpp
#	wallet/core/common.h
#	wallet/core/simple_transaction.cpp
#	wallet/transactions/swaps/CMakeLists.txt
#	wallet/transactions/swaps/swap_transaction.cpp
#	wallet/unittests/CMakeLists.txt
#	wallet/unittests/wallet_test.cpp

+++ p2p/unittest/twopeers_test.cpp
@@ -52,7 +52,7 @@ struct Response {
 struct MessageHandler : IErrorHandler {
 
     void on_protocol_error(uint64_t fromStream, ProtocolError error) override {
-        cout << __FUNCTION__ << "(" << fromStream << "," << error << ")" << endl;
+        cout << __FUNCTION__ << "(" << fromStream << "," << static_cast<int32_t>(error) << ")" << endl;
 		Reactor::get_Current().stop();
     }
